requirement.txt
fastapi==0.110.0
uvicorn==0.25.0
sqlalchemy==2.0.25
pandas==2.1.4
llama-cpp-python==0.2.27
matplotlib==3.8.2
python-multipart==0.0.6
requests==2.31.0
pydantic==2.6.1
python-dotenv==1.0.0